Fix Site class length lint and flaky tab spec
Extract robots methods from Site into SiteRobots concern to bring
the class under the 190-line RuboCop limit (was 192 after slug changes).

Fix flaky partner_save_bar_spec by checking radio :checked state
instead of CSS panel visibility in tab helpers — avoids CI timing
issues with daisyUI's :checked rendering.
